修复剧本试玩时自由对话轮数/结束语提示不显示：搭新场提前到引擎启动前
